Created a new package (lesson_1), moved to package 'lesson_1' all other packages, classes and interfaces.
