--- src/main/java/ru/knyazevvb/Obstacles/Barrier.java
@@ -1,6 +1,0 @@
-package ru.knyazevvb.Obstacles;
-import ru.knyazevvb.Players.*;
-
-public interface Barrier {
-    boolean doBarrier(Player player);
-}
