+++ src/main/java/ru/knyazevvb/lesson_1/Obstacles/Barrier.java
@@ -1,0 +1,6 @@
+package ru.knyazevvb.lesson_1.Obstacles;
+import ru.knyazevvb.lesson_1.Players.*;
+
+public interface Barrier {
+    boolean doBarrier(Player player);
+}
