+++ src/main/java/ru/knyazevvb/lesson_1/Obstacles/Treadmill.java
@@ -1,6 +1,6 @@
-package ru.knyazevvb.Obstacles;
+package ru.knyazevvb.lesson_1.Obstacles;
 
-import ru.knyazevvb.Players.*;
+import ru.knyazevvb.lesson_1.Players.*;
 
 public class Treadmill implements Barrier{
     private double length;
