+++ src/main/java/ru/knyazevvb/lesson_1/Players/Player.java
@@ -1,4 +1,4 @@
-package ru.knyazevvb.Players;
+package ru.knyazevvb.lesson_1.Players;
 
 public interface Player {
     boolean run(double runLength);
